requirements.txt
GitPython 
Resipy
pyvistaqt
